pyproject.toml
[project]
name = "retail-sage"
version = "0.1.0"
description = "Semantic Analytics & Governed Execution — AI-Powered Retail Analytics Agent"
requires-python = ">=3.11"
dependencies = [
    "duckdb>=1.1.0",
    "dbt-core>=1.8.0",
    "dbt-duckdb>=1.8.0",
    "chromadb>=0.5.0",
    "anthropic>=0.40.0",
    "streamlit>=1.38.0",
    "pandas>=2.2.0",
    "pyarrow>=17.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent"]

[tool.ruff]
target-version = "py311"
line-length = 100
